Add in autoload for the pivx conf file

+++ install.sh
@@ -1,21 +1,38 @@
-#!/bin/bash
+# !/bin/bash
 # The goal here is to create a .env file with the required enviornment varibles to connect to the pivx daemon.
 # In the future this will also be used to download and install a daemon if one doesn't exist
 
-#Create the env file
+# Create the env file
 if [ -e .env ]
 then
     echo "Environment varibles already created, delete or edit the .env file to change the configuration"
 else
+    # Check if a pivx config file exists
     file=".env"
-    #RPC SETTINGS
-    read -p "RPC wallet Port: " rpcPort
-    echo "WALLET_PORT="$rpcPort > $file
-    read -p "RPC wallet user: " rpcUser
-    echo "WALLET_USER="$rpcUser >> $file
-    read -p "RPC wallet Password: " rpcPassword
-    echo "WALLET_PASSWORD="$rpcPassword >> $file
-    #EMAIL SETTINGS
+    if [ -f "${input/\~/$HOME/.pivx/pivx.conf}"]
+    then
+    echo "pivx config file exists"
+    # Ask to load if a user would like to load the information
+    read -p "Would you like us to autoload the pivx config file?: " autoloadpivxconf
+    case $autoloadpivxconf in
+        [yY]) echo "Reading from file";
+            homedir=$( getent passwd "$USER" | cut -d: -f6 )
+            source "$homedir/.pivx/pivx.conf";
+            echo "WALLET_PORT="$rpcport > $file;
+            echo "WALLET_USER="$rpcuser >> $file;
+            echo "WALLET_PASSWORD="$rpcpassword >> $file;
+            echo "loaded configs";;
+        [nN])     
+            # RPC SETTINGS
+            read -p "RPC wallet Port: " rpcPort;
+            echo "WALLET_PORT="$rpcPort > $file;
+            read -p "RPC wallet user: " rpcUser;
+            echo "WALLET_USER="$rpcUser >> $file;
+            read -p "RPC wallet Password: " rpcPassword;
+            echo "WALLET_PASSWORD="$rpcPassword >> $file;;
+    esac
+    fi
+    # EMAIL SETTINGS
     read -p "Do you want to set up SMTP? (Y/n): " smtpyesno
     case $smtpyesno in
         [yY]) echo "SMTP_YES_NO=t" >> $file;
@@ -39,28 +56,28 @@ else
 
         [nN]) echo "SMTP_YES_NO=f">> $file;; 
     esac
-    #ADMIN PANNEL SETTINGS
+    # ADMIN PANNEL SETTINGS
     read -p "Do you want the node to connect to or create an admin panel (cnc module)? (Y/n): " cncyesno
     case $cncyesno in
-        #Ask if we are going to create the admin panel with this node
+        # Ask if we are going to create the admin panel with this node
         [yY]) echo "adminPanelSystemEnabled=t" >> $file;
             read -p "Use this node as an admin panel? (y/n)" : adminpanel;
             echo "ADMIN_PANEL: " adminpanel
             if [adminpanel == y]
             then
-                #We are using this node as the admin backend
-                #If so make sure everything is port forwarded correctly or has an open firewall
+                # We are using this node as the admin backend
+                # If so make sure everything is port forwarded correctly or has an open firewall
                 read -p "Please verify that you have opened the approprate ports and this node is connectable" portsAreOpen
-                #Make the user create a username and password and store that
+                # Make the user create a username and password and store that
                 read -p "Create a username that you wish to use :" : adminPanelUserName
                 echo "ADMIN_PANEL_USER=" adminPanelUserName
                 read -p "Create a password that you wish to use :" : adminPanelPassword
                 echo "ADMIN_PANEL_PASSWORD=" adminPanelPassword
-                #We need to update these on the first login to store them more securly and remove them from here
-                #This is mainly just to allow for ease of use
+                # We need to update these on the first login to store them more securly and remove them from here
+                # This is mainly just to allow for ease of use
             else
-                #We aren't using this node as the admin backend
-                #If not then ask for the ip or url to the system that is running the admin panel and user/pass
+                # We aren't using this node as the admin backend
+                # If not then ask for the ip or url to the system that is running the admin panel and user/pass
                 read -p "Enter the IP address of the admin panel:" adminPanelIpAdress
                 echo "ADMIN_PANEL_IP=" adminPanelIpAdress
                 read -p "Enter the API key for this node, you can create that in admin panel" adminPanelAPINode
@@ -68,8 +85,8 @@ else
             fi
             
     esac
-    #Download bootstrap it gets stuck or forked
-    #Probably will do this if it won't move within x amount of minutes (defaulting to an 30 minutes or 30 blocks)
+    # Download bootstrap it gets stuck or forked
+    # Probably will do this if it won't move within x amount of minutes (defaulting to an 30 minutes or 30 blocks)
     read -p "Download and load bootstrap if the wallet gets stuck: " bootstrap
     case $bootstrap in 
         [yY]) echo "BOOTSTRAP=t" >> $file;
@@ -78,14 +95,14 @@ else
         [nN]) echo "BOOTSTRAP=f">> $file;;
         *) echo "invalid response defaulting to false";;
     esac
-    #Ask if we want to bootstrap if forked
+    # Ask if we want to bootstrap if forked
     read -p "Download and load bootstrap if wallet is forked: (Y/n) " bootstrapFork
     case $bootstrapFork in
         [yY]) echo "BOOTSTRAP_FORK=t" >> $file;;
         *) echo "BOOTSTRAP_FORK=f" >> $file;;
     esac
 
-    #DAEMON SETTINGS
+    # DAEMON SETTINGS
     read -p "Restart Wallet every 15 minutes?: (Y/n) " restartWallet
     case $restartWallet in 
         [yY]) echo "RESTART_WALLET=t" >> $file;
@@ -97,5 +114,5 @@ else
     cat $file
 fi
 
-#Start the node module
+# Start the node module
 node index.js
